Preserve original order for equivalent items in asynch merge sort
When two items compare as equal, the sort should preserve the original order of the items relative to each other.  In other words, if item A and item B are deemed equivalent by the sort comparator, and item A appeared before item B in the original unsorted list, then item A should continue to appear before item B in the sorted list.  This makes for a more predictable stable ordering.